Move geoserver_layername_to_uuid to layer.geoserver package
